fix(app):decimal_score

+++ database/migrations/2024_06_06_212338_competencies_tests_schema.php
@@ -128,7 +128,7 @@ return new class extends Migration
             $table->foreignId(self::TEST_ID)->constrained()->onDelete('cascade');
             $table->foreignId(self::USER_ID)->constrained()->onDelete('cascade');
             $table->foreignId(self::STATUS_ID)->constrained('statuses')->onDelete('cascade');
-            $table->integer('score')->default(0);
+            $table->float('score')->default(0);
             $table->timestamps();
         });
     }
